move RemoveStrategy to LoadStrategyScreen
