+++ src/components/screen/strategies.rs
@@ -50,11 +50,6 @@ impl StrategiesScreenCommands {
                     description: "Create and edit",
                     key_type: KeyType::Command,
                 },
-                CommandPalletKey {
-                    key: 'd',
-                    description: "Delete",
-                    key_type: KeyType::Command,
-                },
             ]),
         }
     }
@@ -75,10 +70,6 @@ impl Component<Message, NoUserEvent> for StrategiesScreenCommands {
                 code: Key::Char('c'),
                 modifiers: KeyModifiers::NONE,
             }) => Some(Message::NextScreen(Screen::LoadStrategy)),
-            Event::Keyboard(KeyEvent {
-                code: Key::Char('d'),
-                modifiers: KeyModifiers::NONE,
-            }) => Some(Message::ExpectingInput(DeleteStrategy)),
             _ => None,
         }
     }
@@ -140,73 +131,6 @@ impl Component<Message, NoUserEvent> for StrategySelect {
                 code: Key::Enter, ..
             }) => {
                 Some(Message::SelectedStrategy(self.selected_index))
-            }
-            Event::Keyboard(KeyEvent {
-                code: Key::Char('q'), ..
-            }) => {
-                Some(Message::ReloadScreen)
-            }
-            _ => None,
-        }
-    }
-}
-
-#[derive(MockComponent)]
-pub(crate) struct DeleteStrategyStruct {
-    component: List,
-    selected_index: usize,
-}
-
-impl DeleteStrategyStruct {
-    pub(crate) fn new(app_state: &AppState) -> Self {
-        let strategies = &app_state.available_strategies;
-                
-        let mut rows = TableBuilder::default();
-        for (name, _) in strategies.iter() {
-            rows.add_col(TextSpan::from(name));
-            rows.add_row();
-        }
-
-        Self {
-            component: List::default()
-                    .title("Select a Strategy. Press 'q' to go back.", Alignment::Center)
-                    .scroll(true)
-                    .highlighted_str("> ")
-                    .rewind(true)
-                    .step(1)
-                    .rows(rows.build())
-                    .selected_line(0),
-                selected_index: 0,
-        }
-    }
-}
-
-impl Component<Message, NoUserEvent> for DeleteStrategyStruct {
-    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Message> {
-        match ev {
-            Event::Keyboard(KeyEvent {
-                code: Key::Down, ..
-            }) => {
-                let max_index = self.component.states.list_len-2;
-                if self.selected_index < max_index {
-                    self.selected_index = self.selected_index + 1;
-                    self.perform(Cmd::Move(Direction::Down));
-                }
-                Some(Message::Redraw)
-            },
-            Event::Keyboard(KeyEvent { 
-                code: Key::Up, .. 
-            }) => {
-                if self.selected_index > 0 {
-                    self.selected_index -= 1;
-                    self.perform(Cmd::Move(Direction::Up));
-                }            
-                Some(Message::Redraw)
-            },
-            Event::Keyboard(KeyEvent {
-                code: Key::Enter, ..
-            }) => {
-                Some(Message::DeleteStrategy(self.selected_index))
             }
             Event::Keyboard(KeyEvent {
                 code: Key::Char('q'), ..
