+++ src/components/screen/strategy_load.rs
@@ -4,7 +4,8 @@ use tui_realm_stdlib::{Paragraph, List, Input};
 use tuirealm::{MockComponent, Component, NoUserEvent, Event, event::{KeyEvent, Key, KeyModifiers}, props::{TextSpan, TableBuilder, Alignment}, command::{Cmd, Direction, CmdResult}, State, StateValue};
 
 use crate::{app::{Message, state::AppState, Screen, strategies::default_strategy_details}, mock_components::{CommandPallet, CommandPalletKey, KeyType, key_event_to_cmd}};
-use crate::app::InputType::{RenameStrategy, LoadStrategy};
+use crate::app::InputType::{RenameStrategy, LoadStrategy, DeleteStrategy};
+
 
 #[derive(MockComponent)]
 pub(crate) struct LoadStrategyScreen {
@@ -71,13 +72,18 @@ impl LoadStrategyScreenCommands {
                         key_type: KeyType::Command,
                     },
                     CommandPalletKey {
-                        key: 'd',
-                        description: "Duplicate",
+                        key: 'c',
+                        description: "Clone",
                         key_type: KeyType::Command,
                     },
                     CommandPalletKey {
-                        key: 'c',
-                        description: "Create",
+                        key: 'd',
+                        description: "Delete",
+                        key_type: KeyType::Command,
+                    },
+                    CommandPalletKey {
+                        key: 'n',
+                        description: "New",
                         key_type: KeyType::Command,
                     },
                     CommandPalletKey {
@@ -97,8 +103,8 @@ impl LoadStrategyScreenCommands {
                         key_type: KeyType::Command,
                     },
                     CommandPalletKey {
-                        key: 'c',
-                        description: "Create",
+                        key: 'n',
+                        description: "New",
                         key_type: KeyType::Command,
                     },
                     CommandPalletKey {
@@ -130,15 +136,19 @@ impl Component<Message, NoUserEvent> for LoadStrategyScreenCommands {
                     modifiers: KeyModifiers::NONE,
                 }) => Some(Message::ExpectingInput(RenameStrategy)),
                 Event::Keyboard(KeyEvent {
-                    code: Key::Char('d'),
+                    code: Key::Char('c'),
                     modifiers: KeyModifiers::NONE,
                 }) => Some(Message::DuplicateStrategy),
+                Event::Keyboard(KeyEvent {
+                    code: Key::Char('d'),
+                    modifiers: KeyModifiers::NONE,
+                }) => Some(Message::ExpectingInput(DeleteStrategy)),    
                 Event::Keyboard(KeyEvent {
                     code: Key::Char('l'),
                     modifiers: KeyModifiers::NONE,
                 }) => Some(Message::ExpectingInput(LoadStrategy)),
                 Event::Keyboard(KeyEvent {
-                    code: Key::Char('c'),
+                    code: Key::Char('n'),
                     modifiers: KeyModifiers::NONE,
                 }) => Some(Message::AddNewStrategy),
                 _ => None,
@@ -154,7 +164,7 @@ impl Component<Message, NoUserEvent> for LoadStrategyScreenCommands {
                     modifiers: KeyModifiers::NONE,
                 }) => Some(Message::ExpectingInput(LoadStrategy)),
                 Event::Keyboard(KeyEvent {
-                    code: Key::Char('c'),
+                    code: Key::Char('n'),
                     modifiers: KeyModifiers::NONE,
                 }) => Some(Message::AddNewStrategy),
                 _ => None,
@@ -263,6 +273,73 @@ impl Component<Message, NoUserEvent> for RenameStrategyStruct {
                     CmdResult::Submit(State::None) => Some(Message::ReloadScreen),
                     _ => Some(Message::Redraw),
                 }
+            }
+            _ => None,
+        }
+    }
+}
+
+#[derive(MockComponent)]
+pub(crate) struct DeleteStrategyStruct {
+    component: List,
+    selected_index: usize,
+}
+
+impl DeleteStrategyStruct {
+    pub(crate) fn new(app_state: &AppState) -> Self {
+        let strategies = &app_state.available_strategies;
+                
+        let mut rows = TableBuilder::default();
+        for (name, _) in strategies.iter() {
+            rows.add_col(TextSpan::from(name));
+            rows.add_row();
+        }
+
+        Self {
+            component: List::default()
+                    .title("Select a Strategy. Press 'q' to go back.", Alignment::Center)
+                    .scroll(true)
+                    .highlighted_str("> ")
+                    .rewind(true)
+                    .step(1)
+                    .rows(rows.build())
+                    .selected_line(0),
+                selected_index: 0,
+        }
+    }
+}
+
+impl Component<Message, NoUserEvent> for DeleteStrategyStruct {
+    fn on(&mut self, ev: Event<NoUserEvent>) -> Option<Message> {
+        match ev {
+            Event::Keyboard(KeyEvent {
+                code: Key::Down, ..
+            }) => {
+                let max_index = self.component.states.list_len-2;
+                if self.selected_index < max_index {
+                    self.selected_index = self.selected_index + 1;
+                    self.perform(Cmd::Move(Direction::Down));
+                }
+                Some(Message::Redraw)
+            },
+            Event::Keyboard(KeyEvent { 
+                code: Key::Up, .. 
+            }) => {
+                if self.selected_index > 0 {
+                    self.selected_index -= 1;
+                    self.perform(Cmd::Move(Direction::Up));
+                }            
+                Some(Message::Redraw)
+            },
+            Event::Keyboard(KeyEvent {
+                code: Key::Enter, ..
+            }) => {
+                Some(Message::DeleteStrategy(self.selected_index))
+            }
+            Event::Keyboard(KeyEvent {
+                code: Key::Char('q'), ..
+            }) => {
+                Some(Message::ReloadScreen)
             }
             _ => None,
         }
